Revert "Task/refactoring"
