+++ app/src/main/java/com/mahmoudelshamy/nytimes/common/Resource.kt
@@ -1,0 +1,7 @@
+package com.mahmoudelshamy.nytimes.common
+
+sealed class Resource<T> {
+    data class Success<T>(val data: T) : Resource<T>()
+    data class Error<T>(val error: AppError, val data: T? = null) : Resource<T>()
+    data class Loading<T>(val data: T? = null) : Resource<T>()
+}
