+++ app/src/main/java/com/mahmoudelshamy/nytimes/common/exceptions/ApiExceptions.kt
@@ -1,4 +1,4 @@
-package com.mahmoudelshamy.nytimes.features.common.domain.exceptions
+package com.mahmoudelshamy.nytimes.common.exceptions
 
 data class ErrorResponseException(val statusCode: Int, override val message: String = "") : Throwable()
