+++ app/src/main/java/com/mahmoudelshamy/nytimes/core/BaseUseCase.kt
@@ -1,28 +1,28 @@
-package com.mahmoudelshamy.nytimes.features.common.domain.base
+package com.mahmoudelshamy.nytimes.core
 
-import com.mahmoudelshamy.nytimes.features.common.domain.AppError
-import com.mahmoudelshamy.nytimes.features.common.domain.Result
-import com.mahmoudelshamy.nytimes.features.common.domain.exceptions.ErrorResponseException
+import com.mahmoudelshamy.nytimes.common.AppError
+import com.mahmoudelshamy.nytimes.common.Resource
+import com.mahmoudelshamy.nytimes.common.exceptions.ErrorResponseException
 import kotlinx.coroutines.flow.FlowCollector
 import java.io.IOException
 import kotlin.coroutines.cancellation.CancellationException
 
 abstract class BaseUseCase {
 
-    protected suspend inline fun <F> FlowCollector<Result<F>>.tryFlow(func: (FlowCollector<Result<F>>) -> Unit) {
+    protected suspend inline fun <F> FlowCollector<Resource<F>>.tryFlow(func: (FlowCollector<Resource<F>>) -> Unit) {
         try {
             func.invoke(this)
         } catch (e: CancellationException) {
             throw e
         } catch (e: IOException) {
             emit(
-                Result.Error(
+                Resource.Error(
                     error = AppError.NetworkError
                 )
             )
         } catch (e: ErrorResponseException) {
             emit(
-                Result.Error(
+                Resource.Error(
                     error = AppError.ApiErrorMessage(
                         statusCode = e.statusCode,
                         message = e.message
@@ -31,7 +31,7 @@ abstract class BaseUseCase {
             )
         } catch (e: Throwable) {
             emit(
-                Result.Error(
+                Resource.Error(
                     error = AppError.GeneralError
                 )
             )
