+++ app/src/main/java/com/mahmoudelshamy/nytimes/features/articles/domain/useCases/GetMostPopularArticlesUseCase.kt
@@ -1,7 +1,7 @@
 package com.mahmoudelshamy.nytimes.features.articles.domain.useCases
 
-import com.mahmoudelshamy.nytimes.features.common.domain.Result
-import com.mahmoudelshamy.nytimes.features.common.domain.base.BaseUseCase
+import com.mahmoudelshamy.nytimes.common.Resource
+import com.mahmoudelshamy.nytimes.core.BaseUseCase
 import com.mahmoudelshamy.nytimes.features.articles.domain.models.Article
 import com.mahmoudelshamy.nytimes.features.articles.domain.repository.ArticlesRepository
 import kotlinx.coroutines.flow.Flow
@@ -12,17 +12,17 @@ class GetMostPopularArticlesUseCase @Inject constructor(
     private val articlesRepository: ArticlesRepository
 ) : BaseUseCase() {
 
-    operator fun invoke(): Flow<Result<List<Article>>> = flow {
+    operator fun invoke(): Flow<Resource<List<Article>>> = flow {
         tryFlow {
             // Loading
             emit(
-                Result.Loading()
+                Resource.Loading()
             )
 
             // Get articles
             val articles = articlesRepository.getMostPopularArticles()
             emit(
-                Result.Success(
+                Resource.Success(
                     data = articles
                 )
             )
